+++ src/main/java/de/samply/teiler/TeilerApplication.java
@@ -5,12 +5,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
-//@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 @SpringBootApplication
 @ComponentScan(basePackages = {"de.samply"})
 @EntityScan(basePackages = {"de.samply.db.model"})
 @EnableJpaRepositories("de.samply.db.repository")
+@EnableScheduling
 public class TeilerApplication {
 
   public static void main(String[] args) {